# C#: save allocations by looking up known method names