Still WIP, but most of the implementation is there.

- server maintains a dictionary with binary blob keys and whenever a method of known name is requested, a cached `string` instance with the name is returned.