@mgravell 2019-07-25
Nice; note that re your `Equals` method, there is a `SequenceEqual` extension method on `MemoryExtensions`  that knows how to compare spans efficiently, in particular note that if your data is bitwise equatable (which yours is) it uses a vectorized (SIMD) implementation when available, falling back to using word-length comparisons (so: 8 bytes at a time on x64), just mopping up the last few bytes the slow way.

The hash-code *could* also be computed with a similar approach, although I don't know whether a helper method exists for that.