@jtattermusch 2019-07-25
> Nice; note that re your `Equals` method, there is a `SequenceEqual` extension method on `MemoryExtensions` that knows how to compare spans efficiently, in particular note that if your data is bitwise equatable (which yours is) it uses a vectorized (SIMD) implementation when available, falling back to using word-length comparisons (so: 8 bytes at a time on x64), just mopping up the last few bytes the slow way. Recommendation: use `MemoryExtensions.SequenceEqual` for your `Equals`:
> 
> ```cs
> public bool Equals(BinaryKey other)
> {
>     return GetSpan().SequenceEqual(other.GetSpan());
> }
> ```
> 
> The hash-code _could_ also be computed with a similar approach, although I don't know whether a helper method exists for that.
> 
> Citations:
> 
> [MemoryExtensions.SequenceEqual](https://github.com/dotnet/corefx/blob/b86783ef9525f22b0576278939264897464f9bbd/src/Common/src/CoreLib/System/MemoryExtensions.cs#L200)
> 
> [SpanHelpers.SequenceEqual](https://github.com/dotnet/corefx/blob/b86783ef9525f22b0576278939264897464f9bbd/src/Common/src/CoreLib/System/SpanHelpers.Byte.cs)
> 
> As a slight caveat to that: `System.Numerics.Vectors` (the SIMD lib) is ... glitchy re assembly binding on .NET Framework. It is great (meaning: not glitchy) on .NET Core. Maybe just use the SIMD version on .NET Core, using a custom `long`-based version otherwise? You can `Cast<,>` spans to different primitives for this purpose.

Thanks for the suggestion, I was pretty sure something like this existed and wanted to investigate, but you beat me to it.