@jtattermusch 2019-08-06
One piece remaining would be to look into saving allocations of the `host` string on the server side. For most servers, the host string will be one out of a few possible values most of the times, but I need to figure out how to cache cache them efficiently (unlike methods, the values are not known in advance).
This can be addressed as a followup PR though.