@jtattermusch 2022-04-29
There seems to be no point of pursuing this further now that https://github.com/grpc/grpc/pull/29225 has been merged.